Fix imports after Priority refactor
