+++ offchain/metadata/parsers/base_parser.py
@@ -3,7 +3,7 @@ from typing import Optional, Protocol
 from offchain.metadata.fetchers.base_fetcher import BaseFetcher
 from offchain.metadata.models.metadata import Metadata
 from offchain.metadata.models.token import Token
-from offchain.metadata.registries.base_registry import Priority
+from offchain.metadata.parsers.priority import Priority
 
 
 class BaseParser(Protocol):
