+++ offchain/metadata/registries/parser_registry.py
@@ -2,8 +2,9 @@ from typing import Optional, Type
 
 from offchain.metadata.parsers.base_parser import BaseParser
 from offchain.metadata.parsers.collection.collection_parser import CollectionParser
+from offchain.metadata.parsers.priority import Priority
 from offchain.metadata.parsers.schema.schema_parser import SchemaParser
-from offchain.metadata.registries.base_registry import BaseRegistry, Priority
+from offchain.metadata.registries.base_registry import BaseRegistry
 
 
 class ParserRegistry(BaseRegistry):
