+++ tests/metadata/registries/test_parser_registry.py
@@ -10,8 +10,8 @@ from offchain.metadata.parsers import (
     UnknownParser,
 )
 from offchain.metadata.parsers.collection.collection_parser import CollectionParser
+from offchain.metadata.parsers.priority import Priority
 from offchain.metadata.parsers.schema.schema_parser import SchemaParser
-from offchain.metadata.registries.base_registry import Priority
 from offchain.metadata.registries.parser_registry import ParserRegistry
